Add TDA8425_emu library
This library aims to emulate the Philips Semiconductors TDA8425 Hi-fi
stereo audio processor.

The original goal is to contribute to the emulation of the AdLib Gold
sound card, which used such integrated circuit to add sound effects to
the audio output, as heard in the beautiful soundtrack of the Dune
videogame, made specifically for that rare sound card.

https://github.com/TexZK/TDA8425_emu