Add float samples to the mixer for AdLib Gold
This gives the user a chance to get rid of the heavy clipping/overflow
in the floppy version of Dune. The game sets up a +15dB bass boost via
the AdLib Gold's stereo processor, which causes the clipping. You could
not get rid of it without this change because the volume scaling in
AddSamples would be performed on the already clipped integer samples.

The CD version works fine without this change because it scales back the
FM volume to 25% not via the mixer volume control, but by generating
samples at lower amplitudes.